cutTrailingZeros prop added to the component
cutTrailingZeros (number) => if passed, then it will look for, and cut, ALL trailing zeros , stopping at cutTrailingZeros decimal point. So cutTrailingZeros  is also used to determine the number of decimal places to leave intact when cutting trailing zeros . It's up to user, currently, to bear in mind the decimalScale value (if used) to not exceed its value. Currently, in case both decimalScale and cutTrailingZero props are used, if cutTrailingZero > decimalScale  and fixedDecimal is used, it will display cutTrailingZero  amount of decimals instead of decimalScale   amount of decimals.

examples where this is usefull:
 if 123,456.780 and cutTrailingZeros = 2 then the result will be => 123,456.78
but if 123,456.780 and  cutTrailingZeros = 3 then the result will be => 123,456.780 .