Remove unnecessary useEffect hook
The affected logic does not need to be wrapped in a useEffect hook for
it to be executed on prop change. Moving this logic outside the hook
also ensures that it runs before the component is rendered, preventing
unnecessary re-renders.

This commit also moves the affected chunk of code so that it follows the
declaration of any values/functions that it uses/calls.